+++ src/parse-entry.js
@@ -1,3 +1,5 @@
+const { prerelease } = require('semver')
+
 exports.parseEntry = entry => {
   const [title, ...other] = entry
     .trim()
@@ -10,8 +12,25 @@ exports.parseEntry = entry => {
   return {
     id: versionNumber,
     date: versionDate,
+    status: computeStatus(versionNumber, title),
     text: other
       .filter(item => !/\[.*\]: http/.test(item))
       .join('\n')
   }
+}
+
+function computeStatus(version, title) {
+  if (prerelease(version)) {
+    return 'prereleased'
+  }
+
+  if (title.match(/\[yanked\]/i)) {
+    return 'yanked'
+  }
+
+  if (title.match(/\[unreleased\]/i)) {
+    return 'unreleased'
+  }
+
+  return 'released'
 }