+++ CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New output properties:
+  - The `version` number of the returned entry
+  - The released `date` of the returned entry
+  - The `status` of the release based on the version number and the title line of the entry.
+    Could be equal to `unreleased`, `prereleased`, `released` or `yanked`.
+    Please refer to https://semver.org/#semantic-versioning-specification-semver for more informations about this.
+
 ### Changed
 - **[BREAKING CHANGE]** If given a specific target version, action will now generate an error response if that version is not found in the changelog.
+- **[BREAKING CHANGE]** `log_entry` output property is renamed to `changes`.
 
 ## [1.3.1] - 2020-07-08
 ### Fixed